Punching Sytem Examples
Here are some examples of stored procedures that I elaborated myself through my employment. Timecard is the punch in and punch out timestamp of the user. Timesheet data is edited by the user to fill in additional details to one's work for a time slot. The final worksheet needs to be sent for approval when the week is completed. I standardized these for educational purposes solely. The queries need to be executed separately by blocks delimited by "#" in this case. You might need to change certain table names, variables and conditions tailored for your needs to fit into your application.